refactor: remove redundant "Provider" from providers/Font
