+++ src/App.tsx
@@ -3,7 +3,7 @@ import { RouterProvider, createBrowserRouter } from "react-router-dom";
 import { Toaster } from "@/components/ui/sonner.tsx";
 import { useTheme } from "@/hooks/useTheme.tsx";
 import Start from "@/pages/Start.tsx";
-import { FontProvider } from "@/providers/FontProvider";
+import { FontProvider } from "@/providers/Font";
 
 const router = createBrowserRouter([
   {
